requirements.txt
Pillow==3.0.0
numpy==1.9.3
scipy==0.16.0
wsgiref==0.1.2
